+++ DataAccessRepository/Entities/PassportInfo.cs
@@ -6,8 +6,8 @@ namespace DataAccessRepository.Entities
 {
     public class PassportInfo
     {
-        public int PassportNumber { get; set; }
-        public int VisaNumber { get; set; }
+        public long PassportNumber { get; set; }
+        public long VisaNumber { get; set; }
         public DateTime VisaExpiryDate { get; set; }
         public DateTime PassportExpiryDate { get; set; }
     }
